move fill_polygon_holes function as a method to its own class
